+++ app/api/auth/routes.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, Body, Depends, HTTPException, Form
 from sqlalchemy.orm import Session
 
 from app import crud, models, schemas
-from app.api import deps
+from app.dependencies import deps
 from app.core import security
 from app.core.config import settings
 from app.core.security import get_password_hash
