+++ app/api/users/routes.py
@@ -6,7 +6,7 @@ from pydantic.networks import EmailStr
 from sqlalchemy.orm import Session
 
 from app import crud, models, schemas
-from app.api import deps
+from app.dependencies import deps
 from app.core import security
 from app.core.config import settings
 from app.utils import send_new_account_email
