+++ app/dependencies/deps.py
@@ -11,7 +11,7 @@ from app.core import security
 from app.core.config import settings
 from app.core.security import decode_access_token
 from app.db.session import SessionLocal
-from .auth_bearer import JWTBearer
+from app.api.auth_bearer import JWTBearer
 
 # reusable_oauth2 = OAuth2PasswordBearer(
 #     tokenUrl=f"{settings.API_V1_STR}/login/access-token"
